Move music into Game (since its the same file for all states) Fix TileLaoder issues (like rock was no collider). Fix the Physicsmanager shutdown bug. ADD ROTATED COLLIDER! readd collider in tiled (now rotated) and move the background decoration elements to the walls so player dont think of it as a grabbable object. Remove unused assets.

+++ src/NLTmxMap.cpp
@@ -184,7 +184,7 @@ NLTmxMap* NLLoadTmxMap(char* xml)
 			const auto rotation_attribute = objectnode->first_attribute("rotation");
 			if (rotation_attribute)
 			{
-				object->rotation = atof(rotation_attribute->value());
+				object->rotation = stof(rotation_attribute->value());
 			}
 
 			const auto properties_node = objectnode->first_node("properties");
